services/core/pyproject.toml
[project]
name = "prostoprobuy-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiobotocore>=2.21.1",
    "aiogram>=3.20.0.post0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "boto3>=1.37.1",
    "celery>=5.5.3",
    "commonmark>=0.9.1",
    "fastapi>=0.115.12",
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "greenlet>=3.2.0",
    "gspread>=6.2.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "json-log-formatter>=1.1.1",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-instrumentation-aiohttp-client>=0.57b0",
    "opentelemetry-instrumentation-celery>=0.57b0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-httpx>=0.57b0",
    "opentelemetry-instrumentation-logging>=0.57b0",
    "opentelemetry-instrumentation-requests>=0.57b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "phonenumbers>=9.0.3",
    "pillow>=11.2.1",
    "pillow-heif>=1.1.0",
    "prometheus-client>=0.22.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.3",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "python-yandex-cloud-logging>=0.1.0",
    "pytz>=2025.2",
    "redis>=6.4.0",
    "ruff>=0.12.1",
    "sqlalchemy>=2.0.40",
    "starlette-exporter>=0.23.0",
    "termcolor>=3.1.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "redis>=6.4.0",
    "celery>=5.5.3",
    "pytz>=2025.2",
    "boto3>=1.37.1",
    "termcolor>=3.1.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "aiobotocore>=2.21.1",
    "aiogram>=3.20.0.post0",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "starlette-exporter>=0.23.0",
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "json-log-formatter>=1.1.1",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-instrumentation-aiohttp-client>=0.57b0",
    "opentelemetry-instrumentation-celery>=0.57b0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-httpx>=0.57b0",
    "opentelemetry-instrumentation-logging>=0.57b0",
    "opentelemetry-instrumentation-requests>=0.57b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "httpx>=0.28.1",
    "phonenumbers>=9.0.3",
    "pillow>=11.2.1",
    "pillow-heif>=1.1.0",
    "prometheus-client>=0.22.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.3",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "python-yandex-cloud-logging>=0.1.0",
    "ruff>=0.12.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.35.0",

]
prod = [
    "redis>=6.4.0",
    "celery>=5.5.3",
    "pytz>=2025.2",
    "boto3>=1.37.1",
    "termcolor>=3.1.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "aiobotocore>=2.21.1",
    "aiogram>=3.20.0.post0",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "starlette-exporter>=0.23.0",
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "json-log-formatter>=1.1.1",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-instrumentation-aiohttp-client>=0.57b0",
    "opentelemetry-instrumentation-celery>=0.57b0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-httpx>=0.57b0",
    "opentelemetry-instrumentation-logging>=0.57b0",
    "opentelemetry-instrumentation-requests>=0.57b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "httpx>=0.28.1",
    "phonenumbers>=9.0.3",
    "pillow>=11.2.1",
    "pillow-heif>=1.1.0",
    "prometheus-client>=0.22.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.3",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "python-yandex-cloud-logging>=0.1.0",
    "ruff>=0.12.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.35.0",
]

[tool.mypy]
strict = true
